Implements complete support for lossless readers.
The writer waits for all the readers to catch up when needed; this is the case
when a the circular buffer is not large enough related to the throughput of the
writer and the readers. If the readers have slower throughput than the write and
the circular buffer is small enough to get to a point where the writer is
len(buf) ahead of the reader(s), it'll hang there until the readers can cope
with the throughput.

Add examples to README.md and examples in godoc. Improve API documentation.

Tested to work under go test -race. Two test cases need fixing.

Increased coverage to 99.2%. The last line is a 32 bits overflow check.